move categorize_time_of_day to time_helpers, update refs
